Move code into ModifiedTripBuilder
